+++ test/HAaveProtocolV2_repay.test.js
@@ -84,7 +84,7 @@ contract('Aave V2', function([_, user, someone]) {
 
   describe('Repay Stable Rate', function() {
     var depositAmount = ether('10000');
-    const borrowAmount = ether('2');
+    const borrowAmount = ether('1');
     const borrowTokenAddr = WETH_TOKEN;
     const borrowTokenProvider = WETH_PROVIDER;
     const rateMode = AAVE_RATEMODE.STABLE;
@@ -389,7 +389,7 @@ contract('Aave V2', function([_, user, someone]) {
 
   describe('Repay Variable Rate', function() {
     var depositAmount = ether('10000');
-    const borrowAmount = ether('2');
+    const borrowAmount = ether('1');
     const borrowTokenAddr = WETH_TOKEN;
     const borrowTokenProvider = WETH_PROVIDER;
     const rateMode = AAVE_RATEMODE.VARIABLE;