+++ test/HBProtocol.test.js
@@ -17,8 +17,8 @@ const { ZERO_ADDRESS } = constants;
 const { expect } = require('chai');
 
 const {
-  BAT_TOKEN,
-  BAT_PROVIDER,
+  LINK_TOKEN,
+  LINK_PROVIDER,
   DAI_TOKEN,
   DAI_PROVIDER,
   B_CDP_MANAGER,
@@ -28,10 +28,7 @@ const {
   MAKER_MCD_JUG,
   MAKER_MCD_VAT,
   MAKER_MCD_JOIN_ETH_A,
-  MAKER_MCD_JOIN_BAT_A,
-  MAKER_MCD_JOIN_USDC_A,
-  MAKER_MCD_JOIN_WBTC_A,
-  MAKER_MCD_JOIN_KNC_A,
+  MAKER_MCD_JOIN_LINK_A,
   MAKER_MCD_JOIN_DAI,
 } = require('./utils/constants');
 const {
@@ -94,8 +91,12 @@ async function approveCdp(cdp, owner, user) {
 
 contract('BProtocol', function([_, user1, user2, someone]) {
   let id;
-  const tokenAddress = BAT_TOKEN;
-  const providerAddress = BAT_PROVIDER;
+  const tokenAddress = LINK_TOKEN;
+  const providerAddress = LINK_PROVIDER;
+  const makerMcdJoinETH = MAKER_MCD_JOIN_ETH_A;
+  const makerMcdJoinETHName = 'ETH-A';
+  const makerMcdJoinToken = MAKER_MCD_JOIN_LINK_A;
+  const makerMcdJoinTokenName = 'LINK-A';
 
   before(async function() {
     this.registry = await Registry.new();
@@ -150,8 +151,10 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           expect(await this.dsRegistry.proxies.call(someone)).eq(ZERO_ADDRESS);
 
           const to = this.hMaker.address;
-          const ilkEth = utils.padRight(utils.asciiToHex('ETH-A'), 64);
-          // const wadD = ether('500');
+          const ilkEth = utils.padRight(
+            utils.asciiToHex(makerMcdJoinETHName),
+            64
+          );
           const [
             generateLimit,
             minCollateral,
@@ -161,7 +164,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           const data = abi.simpleEncode(
             'openLockETHAndDraw(uint256,address,address,bytes32,uint256)',
             value,
-            MAKER_MCD_JOIN_ETH_A,
+            makerMcdJoinETH,
             MAKER_MCD_JOIN_DAI,
             ilkEth,
             wadD
@@ -201,7 +204,10 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           expect(await this.dsRegistry.proxies.call(someone)).eq(ZERO_ADDRESS);
 
           const to = this.hMaker.address;
-          const ilkEth = utils.padRight(utils.asciiToHex('ETH-A'), 64);
+          const ilkEth = utils.padRight(
+            utils.asciiToHex(makerMcdJoinETHName),
+            64
+          );
           const [
             generateLimit,
             minCollateral,
@@ -211,7 +217,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           const data = abi.simpleEncode(
             'openLockETHAndDraw(uint256,address,address,bytes32,uint256)',
             MAX_UINT256,
-            MAKER_MCD_JOIN_ETH_A,
+            makerMcdJoinETH,
             MAKER_MCD_JOIN_DAI,
             ilkEth,
             wadD
@@ -244,7 +250,10 @@ contract('BProtocol', function([_, user1, user2, someone]) {
 
         it('User has proxy', async function() {
           const to = this.hMaker.address;
-          const ilkEth = utils.padRight(utils.asciiToHex('ETH-A'), 64);
+          const ilkEth = utils.padRight(
+            utils.asciiToHex(makerMcdJoinETHName),
+            64
+          );
           const [
             generateLimit,
             minCollateral,
@@ -254,7 +263,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           const data = abi.simpleEncode(
             'openLockETHAndDraw(uint256,address,address,bytes32,uint256)',
             value,
-            MAKER_MCD_JOIN_ETH_A,
+            makerMcdJoinETH,
             MAKER_MCD_JOIN_DAI,
             ilkEth,
             wadD
@@ -286,7 +295,10 @@ contract('BProtocol', function([_, user1, user2, someone]) {
 
         it('User has proxy with max amount', async function() {
           const to = this.hMaker.address;
-          const ilkEth = utils.padRight(utils.asciiToHex('ETH-A'), 64);
+          const ilkEth = utils.padRight(
+            utils.asciiToHex(makerMcdJoinETHName),
+            64
+          );
           const [
             generateLimit,
             minCollateral,
@@ -296,7 +308,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           const data = abi.simpleEncode(
             'openLockETHAndDraw(uint256,address,address,bytes32,uint256)',
             MAX_UINT256,
-            MAKER_MCD_JOIN_ETH_A,
+            makerMcdJoinETH,
             MAKER_MCD_JOIN_DAI,
             ilkEth,
             wadD
@@ -329,9 +341,6 @@ contract('BProtocol', function([_, user1, user2, someone]) {
     });
 
     describe('Lock Token', function() {
-      const tokenAddress = BAT_TOKEN;
-      const providerAddress = BAT_PROVIDER;
-
       describe('Draw Dai', function() {
         it('User does not has proxy', async function() {
           const daiUser = await this.dai.balanceOf.call(someone);
@@ -342,7 +351,10 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           expect(await this.dsRegistry.proxies.call(someone)).eq(ZERO_ADDRESS);
 
           const to = this.hMaker.address;
-          const ilkToken = utils.padRight(utils.asciiToHex('BAT-A'), 64);
+          const ilkToken = utils.padRight(
+            utils.asciiToHex(makerMcdJoinTokenName),
+            64
+          );
           const [
             generateLimit,
             minCollateral,
@@ -351,7 +363,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           const wadC = minCollateral;
           const data = abi.simpleEncode(
             'openLockGemAndDraw(address,address,bytes32,uint256,uint256)',
-            MAKER_MCD_JOIN_BAT_A,
+            makerMcdJoinToken,
             MAKER_MCD_JOIN_DAI,
             ilkToken,
             wadC,
@@ -389,7 +401,10 @@ contract('BProtocol', function([_, user1, user2, someone]) {
 
         it('User has proxy', async function() {
           const to = this.hMaker.address;
-          const ilkToken = utils.padRight(utils.asciiToHex('BAT-A'), 64);
+          const ilkToken = utils.padRight(
+            utils.asciiToHex(makerMcdJoinTokenName),
+            64
+          );
           const [
             generateLimit,
             minCollateral,
@@ -398,7 +413,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           const wadD = generateLimit;
           const data = abi.simpleEncode(
             'openLockGemAndDraw(address,address,bytes32,uint256,uint256)',
-            MAKER_MCD_JOIN_BAT_A,
+            makerMcdJoinToken,
             MAKER_MCD_JOIN_DAI,
             ilkToken,
             wadC,
@@ -446,13 +461,13 @@ contract('BProtocol', function([_, user1, user2, someone]) {
       const new1 = abi.simpleEncode(
         'open(address,bytes32,address)',
         B_CDP_MANAGER,
-        utils.padRight(utils.asciiToHex('ETH-A'), 64),
+        utils.padRight(utils.asciiToHex(makerMcdJoinETHName), 64),
         this.user1Proxy.address
       );
       const new2 = abi.simpleEncode(
         'open(address,bytes32,address)',
         B_CDP_MANAGER,
-        utils.padRight(utils.asciiToHex('BAT-A'), 64),
+        utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64),
         this.user2Proxy.address
       );
       await this.user1Proxy.execute(B_PROXY_ACTIONS, new1, { from: user1 });
@@ -477,7 +492,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
         const data = abi.simpleEncode(
           'safeLockETH(uint256,address,uint256)',
           value,
-          MAKER_MCD_JOIN_ETH_A,
+          makerMcdJoinETH,
           cdp
         );
         const receipt = await this.proxy.execMock(to, data, {
@@ -500,7 +515,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
         const data = abi.simpleEncode(
           'safeLockETH(uint256,address,uint256)',
           MAX_UINT256,
-          MAKER_MCD_JOIN_ETH_A,
+          makerMcdJoinETH,
           cdp
         );
         const receipt = await this.proxy.execMock(to, data, {
@@ -522,9 +537,6 @@ contract('BProtocol', function([_, user1, user2, someone]) {
     describe('Lock Token', function() {
       let balanceUser;
       let tokenUser;
-      const tokenAddress = BAT_TOKEN;
-      const providerAddress = BAT_PROVIDER;
-
       before(async function() {
         cdp = await this.cdpManager.last.call(this.user2Proxy.address);
         expect(cdp).to.be.bignumber.not.eq(new BN('0'));
@@ -542,12 +554,12 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           generateLimit,
           minCollateral,
         ] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('BAT-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64)
         );
         const wad = generateLimit;
         const data = abi.simpleEncode(
           'safeLockGem(address,uint256,uint256)',
-          MAKER_MCD_JOIN_BAT_A,
+          makerMcdJoinToken,
           cdp,
           wad
         );
@@ -571,12 +583,12 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           generateLimit,
           minCollateral,
         ] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('BAT-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64)
         );
         const wad = generateLimit;
         const data = abi.simpleEncode(
           'safeLockGem(address,uint256,uint256)',
-          MAKER_MCD_JOIN_BAT_A,
+          makerMcdJoinToken,
           cdp,
           MAX_UINT256
         );
@@ -610,9 +622,9 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           'openLockETHAndDraw(address,address,address,address,bytes32,uint256)',
           B_CDP_MANAGER,
           MAKER_MCD_JUG,
-          MAKER_MCD_JOIN_ETH_A,
+          makerMcdJoinETH,
           MAKER_MCD_JOIN_DAI,
-          utils.padRight(utils.asciiToHex('ETH-A'), 64),
+          utils.padRight(utils.asciiToHex(makerMcdJoinETHName), 64),
           ether('0')
         );
         await this.user1Proxy.execute(B_PROXY_ACTIONS, new1, {
@@ -631,7 +643,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
         const wad = ether('1');
         const data = abi.simpleEncode(
           'freeETH(address,uint256,uint256)',
-          MAKER_MCD_JOIN_ETH_A,
+          makerMcdJoinETH,
           cdp,
           wad
         );
@@ -652,7 +664,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
         const wad = ether('1');
         const data = abi.simpleEncode(
           'freeETH(address,uint256,uint256)',
-          MAKER_MCD_JOIN_ETH_A,
+          makerMcdJoinETH,
           cdp,
           wad
         );
@@ -671,7 +683,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
         const wad = ether('1');
         const data = abi.simpleEncode(
           'freeETH(address,uint256,uint256)',
-          MAKER_MCD_JOIN_ETH_A,
+          makerMcdJoinETH,
           cdp,
           wad
         );
@@ -695,9 +707,9 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           'openLockGemAndDraw(address,address,address,address,bytes32,uint256,uint256,bool)',
           B_CDP_MANAGER,
           MAKER_MCD_JUG,
-          MAKER_MCD_JOIN_BAT_A,
+          makerMcdJoinToken,
           MAKER_MCD_JOIN_DAI,
-          utils.padRight(utils.asciiToHex('BAT-A'), 64),
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64),
           tokenAmount,
           ether('0'),
           true
@@ -722,7 +734,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
         const wad = ether('100');
         const data = abi.simpleEncode(
           'freeGem(address,uint256,uint256)',
-          MAKER_MCD_JOIN_BAT_A,
+          makerMcdJoinToken,
           cdp,
           wad
         );
@@ -741,7 +753,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
         const wad = ether('100');
         const data = abi.simpleEncode(
           'freeGem(address,uint256,uint256)',
-          MAKER_MCD_JOIN_BAT_A,
+          makerMcdJoinToken,
           cdp,
           wad
         );
@@ -759,7 +771,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
         const wad = ether('100');
         const data = abi.simpleEncode(
           'freeGem(address,uint256,uint256)',
-          MAKER_MCD_JOIN_BAT_A,
+          makerMcdJoinToken,
           cdp,
           wad
         );
@@ -785,9 +797,9 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           'openLockETHAndDraw(address,address,address,address,bytes32,uint256)',
           B_CDP_MANAGER,
           MAKER_MCD_JUG,
-          MAKER_MCD_JOIN_ETH_A,
+          makerMcdJoinETH,
           MAKER_MCD_JOIN_DAI,
-          utils.padRight(utils.asciiToHex('ETH-A'), 64),
+          utils.padRight(utils.asciiToHex(makerMcdJoinETHName), 64),
           ether('0')
         );
         await this.user1Proxy.execute(B_PROXY_ACTIONS, new1, {
@@ -806,7 +818,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           generateLimit,
           minCollateral,
         ] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('ETH-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinETHName), 64)
         );
         const wad = generateLimit;
         const data = abi.simpleEncode(
@@ -831,7 +843,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           generateLimit,
           minCollateral,
         ] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('BAT-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64)
         );
         const wad = generateLimit;
         const data = abi.simpleEncode(
@@ -855,7 +867,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           generateLimit,
           minCollateral,
         ] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('BAT-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64)
         );
         const wad = generateLimit;
         const data = abi.simpleEncode(
@@ -883,16 +895,16 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           generateLimit,
           minCollateral,
         ] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('BAT-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64)
         );
         const tokenAmount = minCollateral;
         const new2 = abi.simpleEncode(
           'openLockGemAndDraw(address,address,address,address,bytes32,uint256,uint256,bool)',
           B_CDP_MANAGER,
           MAKER_MCD_JUG,
-          MAKER_MCD_JOIN_BAT_A,
+          makerMcdJoinToken,
           MAKER_MCD_JOIN_DAI,
-          utils.padRight(utils.asciiToHex('BAT-A'), 64),
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64),
           tokenAmount,
           ether('0'),
           true
@@ -918,7 +930,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           generateLimit,
           minCollateral,
         ] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('ETH-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinETHName), 64)
         );
         const wad = generateLimit;
         const data = abi.simpleEncode(
@@ -943,7 +955,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           generateLimit,
           minCollateral,
         ] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('BAT-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64)
         );
         const wad = generateLimit;
         const data = abi.simpleEncode(
@@ -967,7 +979,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           generateLimit,
           minCollateral,
         ] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('BAT-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64)
         );
         const wad = generateLimit;
         const data = abi.simpleEncode(
@@ -997,16 +1009,16 @@ contract('BProtocol', function([_, user1, user2, someone]) {
       beforeEach(async function() {
         const etherAmount = ether('10');
         [generateLimit, minCollateral] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('ETH-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinETHName), 64)
         );
         const daiAmount = generateLimit.add(ether('10'));
         const new1 = abi.simpleEncode(
           'openLockETHAndDraw(address,address,address,address,bytes32,uint256)',
           B_CDP_MANAGER,
           MAKER_MCD_JUG,
-          MAKER_MCD_JOIN_ETH_A,
+          makerMcdJoinETH,
           MAKER_MCD_JOIN_DAI,
-          utils.padRight(utils.asciiToHex('ETH-A'), 64),
+          utils.padRight(utils.asciiToHex(makerMcdJoinETHName), 64),
           daiAmount
         );
         await this.user1Proxy.execute(B_PROXY_ACTIONS, new1, {
@@ -1089,7 +1101,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           generateLimit,
           minCollateral,
         ] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('BAT-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64)
         );
         const tokenAmount = minCollateral;
         const daiAmount = generateLimit.add(ether('10'));
@@ -1098,9 +1110,9 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           'openLockGemAndDraw(address,address,address,address,bytes32,uint256,uint256,bool)',
           B_CDP_MANAGER,
           MAKER_MCD_JUG,
-          MAKER_MCD_JOIN_BAT_A,
+          makerMcdJoinToken,
           MAKER_MCD_JOIN_DAI,
-          utils.padRight(utils.asciiToHex('BAT-A'), 64),
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64),
           tokenAmount,
           daiAmount,
           true
@@ -1160,7 +1172,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           generateLimit,
           minCollateral,
         ] = await getGenerateLimitAndMinCollateral(
-          utils.padRight(utils.asciiToHex('BAT-A'), 64)
+          utils.padRight(utils.asciiToHex(makerMcdJoinTokenName), 64)
         );
         const wad = generateLimit;
         await this.dai.transfer(