+++ test/HComptroller.test.js
@@ -94,7 +94,9 @@ contract('Comptroller', function([_, user, someone]) {
     await evmRevert(id);
   });
 
-  describe('Claim COMP', function() {
+  // NOTE: Because COMP Hack event, compound stop to issue $COMP now.
+  // Remove `skip` when compound issue $COMP again.
+  describe.skip('Claim COMP', function() {
     before(async function() {
       await this.comptroller.claimComp(user);
     });